+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### 🐞 Bug fixes
 - Fixes line flickering problem ([#5094](https://github.com/maplibre/maplibre-gl-js/pull/5094))
 - Fix regression in render world copies ([#5101](https://github.com/maplibre/maplibre-gl-js/pull/5101))
+- Fix poor performance in Chrome related to passing matrices to WebGL ([#5072](https://github.com/maplibre/maplibre-gl-js/pull/5072))
 - _...Add new stuff here..._
 
 ## 5.0.0-pre.7